Change Template to not include constructor, modify code template, add Scenario1(WIP), simplify TeleOp by adding setDrivePower instead of two lines.

+++ FtcRobotController/src/main/java/com/qualcomm/ftcrobotcontroller/opmodes/TeleOp.java
@@ -104,7 +104,7 @@ public class TeleOp extends BasicFunctions {
 		left =  (float)scaleInput(left * scalePower);
 
 		// write the values to the motors
-        setDrivePower(left, right);
+        setDrivePower((double)left, (double)right);
 
 		/*
 		 * Send telemetry data back to driver station. Note that if we are using
